requirements.txt
# Web Framework
flask==3.0.3
flask-socketio==5.4.1
python-socketio==5.11.4
flask-cors==5.0.0

# Google Cloud STT
google-cloud-speech==2.28.0

# OpenAI (Whisper + GPT-4 consolidation)
openai==1.58.1

# Anthropic Claude (optional)
anthropic==0.42.0

# Audio Processing
pydub==0.25.1
numpy==2.2.1

# Database
sqlalchemy==2.0.36

# Environment & Config
python-dotenv==1.0.1

# Async Support
aiofiles==24.1.0
aiohttp==3.11.11

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
flake8==7.1.1
